Create HealthProfessional.js
A non Homogneous array with different type of objects. Our motto is to make a function to get the total based on provided criteria. Also it would be possible to filter certain values if the criteria does not match. It can search by profession types and state (e.g. find all who are ready to start the oepration)